+++ podcasts/AppDelegate.swift
@@ -63,6 +63,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             ServerConfig.shared.playbackDelegate = PlaybackManager.shared
             checkDefaults()
 
+            logActiveDownloadTasks()
             logStaleDownloads()
             postLaunchSetup()
             checkIfRestoreCleanupRequired()
@@ -70,6 +71,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             ImageManager.sharedManager.updatePodcastImagesIfRequired()
             WidgetHelper.shared.cleanupAppGroupImages()
             SiriShortcutsManager.shared.setup()
+
+            if FeatureFlag.downloadFixes.enabled {
+                DownloadManager.shared.startAllQueued()
+            }
         }
 
         badgeHelper.setup()